#281 Code refactoring
