+++ src/LightBDD.Core/Execution/Implementation/GlobalSetUp/GlobalResourceSetUp.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using LightBDD.Core.Dependencies;
 
-namespace LightBDD.Core.Execution.Implementation.GlobalInitialization;
+namespace LightBDD.Core.Execution.Implementation.GlobalSetUp;
 
 internal class GlobalResourceSetUp<TDependency> : IGlobalSetUp where TDependency : IGlobalResourceSetUp
 {
