+++ src/LightBDD.Core/Execution/Implementation/GlobalSetUp/IGlobalSetUp.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using LightBDD.Core.Dependencies;
 
-namespace LightBDD.Core.Execution.Implementation.GlobalInitialization;
+namespace LightBDD.Core.Execution.Implementation.GlobalSetUp;
 
 internal interface IGlobalSetUp
 {
